Update float point on android speaker engine (#128)
* update minium android version to 21 for float-point source
* Update float-point source on android speaker engine
* Refactoring with  TxEngine & WaveRenderer  & EuPIRenderer

+++ euphony/src/main/cpp/core/source/EuPIOscillator.cpp
@@ -24,7 +24,7 @@ void Euphony::EuPIOscillator::setWaveOn(bool isWaveOn) {
 void Euphony::EuPIOscillator::renderAudio(float *data, int32_t numFrames) {
     /* mapping frequencies data */
     for(int i = 0; i < numFrames; ++i) {
-        data[i] = (float) (sin(mPhase) * mAmplitude);
+        data[i] = sin(mPhase) * mAmplitude;
         mPhase += mPhaseIncrement;
         if (mPhase > kTwoPi) mPhase -= kTwoPi;
     }
